# docs/swagger.yaml
basePath: /
definitions:
...
  httpx.ErrorDetail:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  httpx.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/httpx.ErrorResponseBody'
      success:
        type: boolean
    type: object
  httpx.ErrorResponseBody:
    properties:
      code:
        type: string
      details:
        items:
          $ref: '#/definitions/httpx.ErrorDetail'
        type: array
      message:
        type: string
    type: object
  httpx.SuccessResponse:
    properties:
      data: {}
      meta: {}
      success:
        type: boolean
    type: object
...
  user.registerReq:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
host: localhost:8080
info:
  contact:
    name: API Support
  description: A REST API for tracking books you've read, want to read, or are currently
    reading.
  termsOfService: http://swagger.io/terms/
  title: Personal Book Tracking API
  version: "1.0"
paths:
...
  /me:
    get:
      consumes:
      - application/json
      description: Get the authenticated user's information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpx.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - Bearer: []
      summary: Get current user
      tags:
      - users
...
  /users/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: Registration request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.registerReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/httpx.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Register a new user
      tags:
      - users
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"